requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
email-validator==2.0.0
phonenumbers==8.13.0
pandas==2.1.0
perplexityai==0.12.0